feat: anchored eval rubric, unconditional grounding clause, half-step scores
Query Lab eval scores had a ceiling effect: 39/40 dimension cells on the
2026-07-08 baseline queries were 5/5. Three changes, verified by re-judging
four known queries (rebaseline 2026-07-10):

1. Anchored judge rubric (query.py _RUBRIC_GROUNDED/_RUBRIC_BLIND, and the
   autopilot self-eval prompt aligned to the same calibration): every 1-5
   level is defined per dimension, and the verdict must name the claim or
   omission behind any score <= 3. Scores now spread 3-5 with diagnosable
   deductions (e.g. ungrounded Ppk thresholds named on query 581).

2. Unconditional grounding clause (prompt_assembler GROUNDING_CLAUSE):
   structured closing instructions demand specific references; when the
   pack lacked one the model was forced to fabricate from memory. The
   say-so-instead escape hatch previously existed only for general_query
   and low-confidence queries; now appended to every intent.

3. Half-step scores (migration 019, int -> float): merged dimension scores
   are the unrounded mean of the two counterbalanced passes (3.5 = passes
   said 3 and 4). Half-up integer rounding hid pass disagreement — a (4,5)
   split displayed as an uncontested 5 while the verdict criticized the
   answer (query 590). Judges still score anchored integers per pass;
   granularity comes from averaging, not a noisier 10-point scale.

Historical eval rows predate the new calibration and are not comparable.

+++ backend/app/models.py
@@ -343,11 +343,14 @@ class EvalScore(Base):
     eval_model: Mapped[str] = mapped_column(String(50), nullable=False)
     answer_mode: Mapped[str] = mapped_column(String(50), nullable=False)  # e.g. "haiku_neuron"
     answer_label: Mapped[str] = mapped_column(String(50), nullable=False)  # e.g. "A"
-    accuracy: Mapped[int] = mapped_column(Integer, nullable=False)      # 1-5
-    completeness: Mapped[int] = mapped_column(Integer, nullable=False)  # 1-5
-    clarity: Mapped[int] = mapped_column(Integer, nullable=False)       # 1-5
-    faithfulness: Mapped[int] = mapped_column(Integer, nullable=False)  # 1-5 (5 = no hallucinations)
-    overall: Mapped[int] = mapped_column(Integer, nullable=False)       # 1-5
+    # 1-5 in half steps: each counterbalanced judge pass scores integer 1-5;
+    # the stored value is the unrounded mean across passes (e.g. 3.5 = one
+    # pass said 3, the other 4). Migration 019.
+    accuracy: Mapped[float] = mapped_column(Float, nullable=False)
+    completeness: Mapped[float] = mapped_column(Float, nullable=False)
+    clarity: Mapped[float] = mapped_column(Float, nullable=False)
+    faithfulness: Mapped[float] = mapped_column(Float, nullable=False)  # 5 = no hallucinations
+    overall: Mapped[float] = mapped_column(Float, nullable=False)
     verdict: Mapped[str | None] = mapped_column(Text, nullable=True)    # free-text verdict
     created_at: Mapped[datetime.datetime] = mapped_column(
         DateTime, server_default=func.now()
